Revert "20150824-02"
This reverts commit 43b44839f6b2a71e02d2e21352c8934d297af914.

+++ SSP_TimerIRQ/syscalls.c
@@ -138,7 +138,7 @@ _getpid (int n)
 
 caddr_t _sbrk (int incr) {
   prev_heap_end = heap_end;
-  if (heap_end + incr > (unsigned int)stack_ptr)
+  if (heap_end + incr > stack_ptr)
     return (caddr_t) -1;
   heap_end += incr;
   return (caddr_t) prev_heap_end;
